requirements.txt
crewai
crewai-tools
python-dotenv
litellm
google-generativeai
streamlit
pypdf
python-docx